% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpxtractor_utils.R
\name{check_file_path}
\alias{check_file_path}
\title{Check the file path.}
\usage{
check_file_path(file)
}
\value{
Error if the file is null or if the file does not exist. Then check if
the extension is .txt
}
\description{
Check the file path.
}
\seealso{
\code{\link{plater}}
}
