% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpxtractor_utils.R
\name{check_that_file_is_non_empty}
\alias{check_that_file_is_non_empty}
\title{Check that file is not empty}
\usage{
check_that_file_is_non_empty(file)
}
\value{
Error if the file is empty.
}
\description{
Check that file is not empty
}
\seealso{
\code{\link{plater}}
}
