% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpxtractor_utils.R
\name{check_that_only_one_file_is_provided}
\alias{check_that_only_one_file_is_provided}
\title{Check that one file is provided.}
\usage{
check_that_only_one_file_is_provided(file)
}
\value{
Error if more than one file is given.
}
\description{
Check that one file is provided.
}
\seealso{
\code{\link{plater}}
}
