% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanfile_to_df.R
\name{cleanfile_to_df}
\alias{cleanfile_to_df}
\title{Make data frame using a clean file.}
\usage{
cleanfile_to_df(cleanfile)
}
\arguments{
\item{cleanfile}{A string}
}
\value{
tidy dataframe
}
\description{
\code{cleanfile_to_df} takes a clean file which is a string type and using
read.table generates a dataframe.
}
\seealso{
\code{\link[utils]{read.table}}
}
