% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_fluorstar_fluorescence_outdata.R
\docType{data}
\name{df_fluorstar_fluorescence_outdata}
\alias{df_fluorstar_fluorescence_outdata}
\title{Full dataset of fluorstar microplate readers.}
\format{
A data frame with 23712 rows and 5 variables:
\describe{
  \item{Wells}{Well id}
  \item{Sample}{Sample X}
  \item{Time}{Time point, in hh:mm:ss format}
  \item{Measurement}{OD600, optical density at 600 nm}
  \item{wavelength}{ wave lenght at which the measurements were performed}
}
}
\source{
\url{http://www.teusinkbruggemanlab.nl/}
}
\usage{
df_fluorstar_fluorescence_outdata
}
\description{
The Measurements of this datasets are absorbance and fluorescence. This for a
8x12 micro plate produced by fluorstar reader. The datasets contain the
Measurement and other attributes for four different timepoints. The dataset
was group by wells and arranged by time. This in order to provide an easy way
to use time series. Both datasets df_absorbance_outdata and
df_fluorscence_outdata have the same structure.
}
\section{Warning}{
 Note that this datsets does not posses the correct time
  step at the begining due to the way that this machines generate the data.
  For example start with 00:00:00 and continue with 00:02:00 when the
  timestep is 5 minutes.
}

\keyword{datasets}
