% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_multiscango_outdata_2.R
\docType{data}
\name{df_multiscango_outdata_2}
\alias{df_multiscango_outdata_2}
\title{Toy dataset of multiscanGO microplate readers.}
\format{
A data frame with 1536 rows and 3 variables:
\describe{
 \item{Wells}{Well id}
 \item{Reading}{Time step, contain the reading number.}
 \item{Measurement}{OD600, optical density at 600 nm}
}
}
\source{
\url{http://www.teusinkbruggemanlab.nl/}
}
\usage{
df_multiscango_outdata_2
}
\description{
Measurements of od600 for a 16x24 micro plate produced by multiscanGO reader.
The dataset contain the Measurement and other attributes for four different
reads. The dataset was grouped by wells and arrange by reading. Both
datatasets df_multiscango_1 and df_multiscango_2 have the same structure.
}
\keyword{datasets}
