% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_time.R
\name{format_time}
\alias{format_time}
\title{Transform the time from hh:mm:ss to hours or minutes}
\usage{
format_time(df_data, time_format = NULL)
}
\arguments{
\item{df_data}{dataframe}

\item{time_format}{which is a character to indicate
whether the the format is in hours or in minutes.}
}
\value{
the input dataframe whit the column Time formatted.
}
\description{
Process column Time of df_data and transform the time in hh:mm:ss format to
hours or minutes.
}
\examples{
#Load example dataframe
data_file <- system.file("extdata",
"test_spectramax_data_1.txt",
package = "mpxtractor")
df_spectramax_outdata_1 <- mpxtractor::read_spectramax_data(data_file)

# format time
df <- mpxtractor::format_time(df_spectramax_outdata_1)

# Time transformed
df
}
