% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_read_multifiles.R
\name{get_input_read_multifiles}
\alias{get_input_read_multifiles}
\title{Get input to give to the function that read multiple files}
\usage{
get_input_read_multifiles(folder = NULL, pattern = NULL, filebyname = NULL)
}
\description{
Get input to give to the function that read multiple files
}
