% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microplate_spectramax_growth_rates.R
\name{microplate_spectramax_gr_plot}
\alias{microplate_spectramax_gr_plot}
\title{Function to plot growth rates from spectramax data.}
\usage{
microplate_spectramax_gr_plot(
  spectramax_data,
  layout_file,
  time = NULL,
  exp_title = NULL,
  windowsize,
  var_to_col
)
}
\arguments{
\item{file}{The path to a proper .csv layout file.}

\item{var_shape}{Assign shape to represent one variable}

\item{var_colour}{Assign colour to other variable}

\item{name_plate_layout}{This argument is optional, add the title.}
}
\value{
Returns a plot that represents the microplate with the given layout.
The plot shows in the x-axis the number of columns and in the y-axis the
letter to identified the row.
}
\description{
This function recive two files which are the layout file .csv and the spectramax
data file .txt.
plot that represents the growth rates into the microplate frame.
}
\section{\code{file} format}{

Note that the .csv layout file needs to be formatted in the correct way. See
examples below.
}

\examples{
file_path <- system.file("extdata", "spectraMax_layout_plate1.csv",
package = "mpxtractor")

# Data is store as a tibble
plot_plate <- plot_layout_file(
   file = file_path, var_shape = "basic", var_colour = "condition",
   name_plate_layout = "My experiment" )

# Show the plot
plot_plate
}
