% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpxtractor.R
\docType{package}
\name{mpxtractor}
\alias{mpxtractor}
\title{mipladxtractor: A package to obtain tidy dataframes from microplates readers
output files.}
\description{
The mipladxtractor provides 4 categories of important functions:
read_spectraMax_data, read_multiscanGo_data, read_fluorStar_data and read_multiple_files.
}
\section{read_spectraMax function}{

The read_spectraMax function process one spectraMax output file at the time..
}

\section{read_multiscanGo_data function}{

The ...
}

\section{read_fluorStar_data function}{

The ....
}

\section{read_multiple_files function}{

The ...
}

