% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout_dataframe.R
\name{plot_layout_dataframe}
\alias{plot_layout_dataframe}
\title{Function to plot layout dataframes}
\usage{
plot_layout_dataframe(platemap_df, var_shape, var_colour, plate_title)
}
\arguments{
\item{platemap_df}{dataframe structure similar to read_layout_files() output}

\item{var_shape}{Assign shape to represent one variable}

\item{var_colour}{Assign colour to other variable}

\item{plate_title}{This argument is optional, add the title.}
}
\value{
Returns a plot that represents the microplate with the given layout.
The plot shows in the x-axis the number of columns and in the y-axis the
letter to identified the row.
}
\description{
This function recive dataframe similar to the output from read_layout_files()
and generate a plot that represents the designed layout of the microplate.
}
\examples{
file_path <- system.file("extdata", "test_spectraMax_layout_1.csv",
  package = "mpxtractor"
)

# Data is store as a tibble
plot_plate <- plot_layout_file(
  file = file_path, var_shape = "basic", var_colour = "condition",
  plate_title = "My experiment"
)

# Show the plot
plot_plate

}
