% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout_file.R
\name{plot_layout_file}
\alias{plot_layout_file}
\title{Function to plot layout files.}
\usage{
plot_layout_file(
  file,
  var_shape = NULL,
  var_colour = NULL,
  add_conc = FALSE,
  plate_title = NULL,
  output_filename
)
}
\arguments{
\item{file}{The path to a proper .csv layout file.}

\item{var_shape}{Assign shape to represent one variable}

\item{var_colour}{Assign colour to other variable}

\item{add_conc}{This argument is optional, TRUE to add concentration.}

\item{plate_title}{This argument is optional, add the title.}

\item{output_filename}{The name of the output file followed by propoer extension, ie. .png}
}
\value{
Returns a plot that represents the microplate with the given layout.
The plot shows in the x-axis the number of columns and in the y-axis the
letter to identified the row.
}
\description{
This function recive one file which is the .csv layout file and generate a
plot that represents the designed layout of the microplate.
}
\section{\code{file} format}{

Note that the .csv layout file needs to be formatted in the correct way. See
files in the examples below.
}

\examples{
file_path <- system.file("extdata", "test_spectraMax_layout_1.csv",
  package = "mpxtractor"
)

# Data is store as a tibble
plot_plate <- plot_layout_file(
  file = file_path, var_shape = "basic", var_colour = "condition", add_conc = TRUE,
  plate_title = "My experiment"
)

# Show the plot
plot_plate

}
