% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout_file.R
\name{plot_lyout_file}
\alias{plot_lyout_file}
\title{Function plot the layout file given in a correct format.}
\usage{
plot_lyout_file(file, var_shape, var_colour, name_plate_layout)
}
\arguments{
\item{file}{The path to a proper .csv layout file.}
}
\value{
Returns a plot that represents the microplate with the given layout.The
plot shows in the x-axis the number of columns and in the y-axis the letter for
identified with the row. The variables are distinguish by shape and by colour.
}
\description{
This function recive one file which is the .csv layout file and generate a
a plot that represents the designed layout of the microplate.
}
\section{\code{file} format}{

Note that the .csv layout file needs to be formatted in the correct way. See
examples below.
}

\examples{
file_path <- system.file("extdata", "spectraMax_layout_plate1.csv",
package = "mpxtractor")

# Data is store as a tibble
plot_plate <- plot_layout_file(
   file = file_path)

# Show the plot
plot_plate
}
