% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiple_data_files.R
\name{read_multiple_data_files}
\alias{read_multiple_data_files}
\title{Multiple plates files}
\usage{
read_multiple_data_files(
  reader_type = NULL,
  dirFiles = NULL,
  file_pattern = NULL,
  filesname = NULL,
  plate_names = NULL
)
}
\description{
Multiple plates files
}
