% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiple_plates.R
\name{read_multiple_plates}
\alias{read_multiple_plates}
\title{Multple plates files}
\usage{
read_multiple_plates(
  dirFiles = NULL,
  file_pattern = NULL,
  filesname = NULL,
  plate_names = NULL
)
}
\description{
Multple plates files
}
