% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiscan_list_output.R
\name{read_multiscan_list_output}
\alias{read_multiscan_list_output}
\title{Function specific to read output files (.txt) from multiscanGO readers machines
 for output format list.}
\usage{
read_multiscan_list_output(file)
}
\arguments{
\item{file}{The path to a proper .txt file formatted by the multiscango machine
for the option list.}
}
\value{
Returns a tibble data frame with all the columns present in the file,
in addition to this one column that contained time in hh:mm:ss format.
}
\description{
This function receive one output file from multiscango plate reader and
generate a tibble dataframe.
}
