% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiscanGO_data.R
\name{read_multiscanGO_data}
\alias{read_multiscanGO_data}
\title{Function specific to read output .txt files from multiscanGO readers.}
\usage{
read_multiscanGO_data(file)
}
\description{
\code{read_multiscanGO_data} returns the data in the .txt as tibble data frame
}
\examples{
file_path <- system.file("extdata", "multiscanGO_1streading.txt", package = "mpxtractor")

# Data is store as a tibble
data <- read_multiscanGO_data(
   file = file_path)

# Now data is tidy
head(data)
}
