% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectraMax_data.R
\name{read_spectraMax_data}
\alias{read_spectraMax_data}
\title{Function specific to read output files from spectraMax readers.}
\usage{
read_spectraMax_data(file)
}
\description{
Function specific to read output files from spectraMax readers.
}
