% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_format_df.R
\name{std_format_df}
\alias{std_format_df}
\title{Transform the dataframe to the standar data frame for microplate data}
\usage{
std_format_df(df)
}
\description{
Transform the dataframe to the standar data frame for microplate data
}
