% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_path.R, R/mpxtractor_utilities.R
\name{check_file_path}
\alias{check_file_path}
\title{CHeck file path}
\usage{
check_file_path(file)

check_file_path(file)
}
\value{
Error if the file is null or if the file does not exist. Then check if
the extension is .txt
}
\description{
CHeck file path

Check the file path.
}
\seealso{
\code{\link{plater}}

\code{\link{plater}}
}
