% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_that_file_is_non_empty.R,
%   R/mpxtractor_utilities.R
\name{check_that_file_is_non_empty}
\alias{check_that_file_is_non_empty}
\title{Check that file is not empty}
\usage{
check_that_file_is_non_empty(file)

check_that_file_is_non_empty(file)
}
\value{
Error if the file is empty.
}
\description{
Check that file is not empty

Check that file is not empty
}
\seealso{
\link{}

\code{\link{plater}}
}
