% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_that_only_one_file_is_provided.R,
%   R/mpxtractor_utilities.R
\name{check_that_only_one_file_is_provided}
\alias{check_that_only_one_file_is_provided}
\title{Check that one file is provided}
\usage{
check_that_only_one_file_is_provided(file)

check_that_only_one_file_is_provided(file)
}
\value{
Error if more than one file is given.
}
\description{
Check that one file is provided

Check that one file is provided.
}
\seealso{
\code{\link{plater}}

\code{\link{plater}}
}
