% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gr_microplate.R
\name{plot_gr_microplate}
\alias{plot_gr_microplate}
\title{Function to plot growth rates over microplate frame.}
\usage{
plot_gr_microplate(
  df_data,
  var_gr,
  exp_title = NULL,
  ws,
  cond_to_col,
  plate_file = NULL
)
}
\arguments{
\item{df_data}{Is a dataframe that combines data files with layout files}

\item{var_gr}{This the attribute to be used to calculate growth rates}

\item{exp_title}{optional, add the title.}

\item{ws}{is the windowsize in hours}

\item{cond_to_col}{The condition from the layout to color}

\item{plate_file}{plate file to be use to compute growth rates in case of multiple files.}
}
\value{
Returns the background plot which is the microplate frame and over this
the plot of growth rates in each well.
The background plot shows in the x-axis the number of columns and in the y-axis
the letter to identified the row.
}
\description{
This function takes a dataframe with the raw data and the information from the
layout file. Calculate growth rate and plot this growth rates over a microplate
frame.
}
\section{Warning}{

Note that the time should be a time series, if the time series is broken the
growth rates are not calculated and the process stop.
}

\examples{
# Get the data file path
file_path_sp <- system.file(
 "extdata",
 "test_gr_spectramax.txt",
 package = "mpxtractor"
)

# Extract the data stored in the files into a df using proper wrangling function
df_sp <- mpxtractor::read_spectramax_data(
 file = file_path_sp
 )
 # get the path to layout file
file_path_layout <- system.file(
 "extdata", "test_spectraMax_layout_1.csv",
 package = "mpxtractor"
)
# combine raw data with layout scheme
df_data_combine <- mpxtractor::combine_data_with_layout(
 df_data = df_sp,
 reader_type = "spectramax",
 layout_files = file_path_layout
)

microplateplot <- mpxtractor::plot_gr_microplate(
 df_data = df_data_combine,
 var_gr = "Measurement",
 exp_title = "Spectramax experiment",
 ws = "2hs",
 cond_to_col = "condition")

#Check vignette **plotting_functions()** for more information.

}
