% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fluorstar_data.R
\name{read_fluorstar_data}
\alias{read_fluorstar_data}
\title{Function specific to read output (.txt) files from fluorStar readers machines.}
\usage{
read_fluorstar_data(file)
}
\description{
\code{read_fluorstar_data} returns the data in the .txt file as tibble
data frame
}
\examples{
file_path <- system.file("extdata", "test_fluorstar_fluorescence.txt",
  package = "mpxtractor"
)

# Data is store as a tibble
data <- read_fluorstar_data(
  file = file_path
)

# Now data is tidy
head(data)

}
