% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_layout_files.R
\name{read_layout_files}
\alias{read_layout_files}
\title{Function specific to read layout files .csv .}
\usage{
read_layout_files(
  data,
  reader_type = NULL,
  dir_lyout_files = NULL,
  layout_file_pattern = NULL,
  layout_files = NULL,
  plate_names = NULL
)
}
\description{
This function recive as argument a dataframe (data) which is given for
one of the functions(read_spectramax_data, read_multiscango_data,
read_fluorstar_data)
}
\details{
\code{read_layout_files} returns the data in the .csv as tibble
data frame
}
\examples{
file_path <- system.file("extdata", "spectraMax_layout_plate1.csv",
package = "mpxtractor")

# Data is store as a tibble
data_layout <- read_layout_files(data, reader_type ="spectramax"
   layout_files = file_path)

# Now data is tidy
head(data)
}
