% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectramax_data.R
\name{read_spectramax_data}
\alias{read_spectramax_data}
\title{Function specific to read output .txt files from spectraMax readers.}
\usage{
read_spectramax_data(file)
}
\description{
\code{read_spectraMax_data} returns the data in the .txt as tibble data frame
}
\examples{
file_path <- system.file("extdata", "spectraMax_1stplate.txt", package = "mpxtractor")

# Data is store as a tibble
data <- read_spectraMax_data(
   file = file_path)

# Now data is tidy
head(data)
}
